create BlockBossLogicProvider;deprecate associated methods from BlockBossLogic
