+++ java/src/jmri/jmrit/whereused/WhereUsedCollectors.java
@@ -1,6 +1,9 @@
 package jmri.jmrit.whereused;
 
+import jmri.jmrit.blockboss.BlockBossLogicProvider;
 import jmri.jmrit.display.EditorManager;
+
+import java.util.Collections;
 import java.util.Enumeration;
 
 import jmri.*;
@@ -80,14 +83,12 @@ public class WhereUsedCollectors {
      */
     static String checkLights(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(LightManager.class).getNamedBeanSet().forEach((light) -> {
-            light.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("LightControl")) {  // NOI18N
-                    String name = light.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                    sb.append(Bundle.getMessage("ReferenceLineData", name, report.usageData));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(LightManager.class).getNamedBeanSet().forEach((light) -> light.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("LightControl")) {  // NOI18N
+                String name = light.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+                sb.append(Bundle.getMessage("ReferenceLineData", name, report.usageData));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceLightControl");  // NOI18N
     }
 
@@ -107,14 +108,12 @@ public class WhereUsedCollectors {
      */
     static String checkRoutes(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(RouteManager.class).getNamedBeanSet().forEach((route) -> {
-            route.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("Route")) {  // NOI18N
-                    String name = route.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                    sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(RouteManager.class).getNamedBeanSet().forEach((route) -> route.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("Route")) {  // NOI18N
+                String name = route.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+                sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceRoutes");  // NOI18N
     }
 
@@ -132,14 +131,12 @@ public class WhereUsedCollectors {
      */
     static String checkBlocks(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(BlockManager.class).getNamedBeanSet().forEach((block) -> {
-            block.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("Block")) {  // NOI18N
-                    String name = block.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                    sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(BlockManager.class).getNamedBeanSet().forEach((block) -> block.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("Block")) {  // NOI18N
+                String name = block.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+                sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceBlock");  // NOI18N
     }
 
@@ -157,14 +154,12 @@ public class WhereUsedCollectors {
      */
     static String checkLayoutBlocks(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(LayoutBlockManager.class).getNamedBeanSet().forEach((layoutBlock) -> {
-            layoutBlock.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("LayoutBlock")) {  // NOI18N
-                    String name = layoutBlock.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                    sb.append(Bundle.getMessage("ReferenceLineData", name, report.usageData));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(LayoutBlockManager.class).getNamedBeanSet().forEach((layoutBlock) -> layoutBlock.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("LayoutBlock")) {  // NOI18N
+                String name = layoutBlock.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+                sb.append(Bundle.getMessage("ReferenceLineData", name, report.usageData));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceLayoutBlock");  // NOI18N
     }
 
@@ -190,7 +185,7 @@ public class WhereUsedCollectors {
      */
     static String checkSignalHeadLogic(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        Enumeration<BlockBossLogic> e = BlockBossLogic.entries();
+        Enumeration<BlockBossLogic> e = Collections.enumeration(InstanceManager.getDefault(BlockBossLogicProvider.class).provideAll());
         while (e.hasMoreElements()) {
             BlockBossLogic ssl = e.nextElement();
             ssl.getUsageReport(bean).forEach((report) -> {
@@ -222,22 +217,20 @@ public class WhereUsedCollectors {
      */
     static String checkSignalMastLogic(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(SignalMastLogicManager.class).getNamedBeanSet().forEach((sml) -> {
-            sml.getUsageReport(bean).forEach((report) -> {
-                String name = bean.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                if (report.usageKey.startsWith("SMLSource")) {  // NOI18N
-                    sb.append(Bundle.getMessage("ReferenceLineData", name, Bundle.getMessage("SourceMast")));  // NOI18N
-                    return;
-                }
-                if (report.usageKey.startsWith("SMLDest")) {  // NOI18N
-                    sb.append(Bundle.getMessage("ReferenceLineData", name, Bundle.getMessage("DestMast")));  // NOI18N
-                    return;
-                }
-                if (report.usageKey.startsWith("SML")) {  // NOI18N
-                    sb.append(Bundle.getMessage("ReferenceLinePair", sml.getSourceMast().getDisplayName(), report.usageBean.getDisplayName()));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(SignalMastLogicManager.class).getNamedBeanSet().forEach((sml) -> sml.getUsageReport(bean).forEach((report) -> {
+            String name = bean.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+            if (report.usageKey.startsWith("SMLSource")) {  // NOI18N
+                sb.append(Bundle.getMessage("ReferenceLineData", name, Bundle.getMessage("SourceMast")));  // NOI18N
+                return;
+            }
+            if (report.usageKey.startsWith("SMLDest")) {  // NOI18N
+                sb.append(Bundle.getMessage("ReferenceLineData", name, Bundle.getMessage("DestMast")));  // NOI18N
+                return;
+            }
+            if (report.usageKey.startsWith("SML")) {  // NOI18N
+                sb.append(Bundle.getMessage("ReferenceLinePair", sml.getSourceMast().getDisplayName(), report.usageBean.getDisplayName()));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceMastSML");  // NOI18N
     }
 
@@ -255,14 +248,12 @@ public class WhereUsedCollectors {
      */
     static String checkSignalGroups(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(SignalGroupManager.class).getNamedBeanSet().forEach((group) -> {
-            group.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("SignalGroup")) {  // NOI18N
-                    String name = group.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                    sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(SignalGroupManager.class).getNamedBeanSet().forEach((group) -> group.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("SignalGroup")) {  // NOI18N
+                String name = group.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+                sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceSignalGroup");  // NOI18N
     }
 
@@ -282,14 +273,12 @@ public class WhereUsedCollectors {
      */
     static String checkOBlocks(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(OBlockManager.class).getNamedBeanSet().forEach((oblock) -> {
-            oblock.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("OBlock")) {  // NOI18N
-                    String name = oblock.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                    sb.append(Bundle.getMessage("ReferenceLineData", name, report.usageData));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(OBlockManager.class).getNamedBeanSet().forEach((oblock) -> oblock.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("OBlock")) {  // NOI18N
+                String name = oblock.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+                sb.append(Bundle.getMessage("ReferenceLineData", name, report.usageData));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceOBlock");  // NOI18N
     }
 
@@ -306,14 +295,12 @@ public class WhereUsedCollectors {
      */
     static String checkWarrants(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(WarrantManager.class).getNamedBeanSet().forEach((warrant) -> {
-            warrant.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("Warrant")) {  // NOI18N
-                    String name = warrant.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                    sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(WarrantManager.class).getNamedBeanSet().forEach((warrant) -> warrant.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("Warrant")) {  // NOI18N
+                String name = warrant.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+                sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceWarrant");  // NOI18N
     }
 
@@ -331,14 +318,12 @@ public class WhereUsedCollectors {
      */
     static String checkEntryExit(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(EntryExitPairs.class).getNamedBeanSet().forEach((destPoint) -> {
-            destPoint.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("EntryExit")) {  // NOI18N
-                    String name = destPoint.getDisplayName();
-                    sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(EntryExitPairs.class).getNamedBeanSet().forEach((destPoint) -> destPoint.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("EntryExit")) {  // NOI18N
+                String name = destPoint.getDisplayName();
+                sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceEntryExit");  // NOI18N
     }
 
@@ -355,15 +340,13 @@ public class WhereUsedCollectors {
      */
     static String checkLogixConditionals(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(LogixManager.class).getNamedBeanSet().forEach((logix) -> {
-            logix.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("ConditionalVariable") || report.usageKey.startsWith("ConditionalAction")) {  // NOI18N
-                    String name = logix.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                    String cdlName = report.usageBean.getDisplayName();
-                    sb.append(Bundle.getMessage("ReferenceLineConditional", name, cdlName, Bundle.getMessage(report.usageKey), report.usageData));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(LogixManager.class).getNamedBeanSet().forEach((logix) -> logix.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("ConditionalVariable") || report.usageKey.startsWith("ConditionalAction")) {  // NOI18N
+                String name = logix.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+                String cdlName = report.usageBean.getDisplayName();
+                sb.append(Bundle.getMessage("ReferenceLineConditional", name, cdlName, Bundle.getMessage(report.usageKey), report.usageData));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceConditionals");  // NOI18N
     }
 
@@ -382,14 +365,12 @@ public class WhereUsedCollectors {
      */
     static String checkSections(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(SectionManager.class).getNamedBeanSet().forEach((section) -> {
-            section.getUsageReport(bean).forEach((report) -> {
-                if (report.usageKey.startsWith("SectionSensor")) {  // NOI18N
-                    String name = section.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                    sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(SectionManager.class).getNamedBeanSet().forEach((section) -> section.getUsageReport(bean).forEach((report) -> {
+            if (report.usageKey.startsWith("SectionSensor")) {  // NOI18N
+                String name = section.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+                sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceSections");  // NOI18N
     }
 
@@ -409,17 +390,15 @@ public class WhereUsedCollectors {
      */
     static String checkTransits(NamedBean bean) {
         StringBuilder sb = new StringBuilder();
-        InstanceManager.getDefault(TransitManager.class).getNamedBeanSet().forEach((transit) -> {
-            transit.getUsageReport(bean).forEach((report) -> {
-                String name = transit.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
-                if (report.usageKey.startsWith("TransitSensor") || report.usageKey.startsWith("TransitSection")) {  // NOI18N
-                    sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
-                }
-                if (report.usageKey.startsWith("TransitAction")) {  // NOI18N
-                    sb.append(Bundle.getMessage("ReferenceLineAction", name, report.usageBean.getDisplayName()));  // NOI18N
-                }
-            });
-        });
+        InstanceManager.getDefault(TransitManager.class).getNamedBeanSet().forEach((transit) -> transit.getUsageReport(bean).forEach((report) -> {
+            String name = transit.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME);
+            if (report.usageKey.startsWith("TransitSensor") || report.usageKey.startsWith("TransitSection")) {  // NOI18N
+                sb.append(Bundle.getMessage("ReferenceLineName", name));  // NOI18N
+            }
+            if (report.usageKey.startsWith("TransitAction")) {  // NOI18N
+                sb.append(Bundle.getMessage("ReferenceLineAction", name, report.usageBean.getDisplayName()));  // NOI18N
+            }
+        }));
         return addHeader(sb, "ReferenceTransits");  // NOI18N
     }
