+++ java/test/jmri/configurexml/LoadAndStoreTestBase.java
@@ -8,7 +8,6 @@ import java.io.InputStreamReader;
 import java.util.Collection;
 import jmri.ConfigureManager;
 import jmri.InstanceManager;
-import jmri.SignalMastLogicManager;
 import jmri.jmrit.logix.WarrantPreferences;
 import jmri.util.FileUtil;
 import jmri.util.JUnitAppender;
@@ -20,7 +19,6 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-import org.picocontainer.behaviors.Stored;
 
 /**
  * Base for testing load-and-store of configuration files.
@@ -341,17 +339,20 @@ public class LoadAndStoreTestBase {
     public void setUp() {
         JUnitUtil.setUp();
         JUnitUtil.resetProfileManager();
+        JUnitUtil.resetInstanceManager();
         JUnitUtil.initConfigureManager();
         JUnitUtil.initInternalTurnoutManager();
         JUnitUtil.initInternalLightManager();
         JUnitUtil.initInternalSensorManager();
+        JUnitUtil.initInternalSignalHeadManager();
         JUnitUtil.initMemoryManager();
+        JUnitUtil.clearBlockBossLogic();
         System.setProperty("jmri.test.no-dialogs", "true");
     }
 
     @After
     public void tearDown() {
-        InstanceManager.getDefault(SignalMastLogicManager.class).dispose();
+        JUnitUtil.closeAllPanels();
         JUnitUtil.clearShutDownManager();
         JUnitUtil.clearBlockBossLogic();
         JUnitUtil.tearDown();
